+++ firmware/application/src/rfid/nfctag/lf/protocols/t55xx.h
@@ -77,6 +77,12 @@ extern "C" {
     T5577_PWD |                   \
     (2 << T5577_MAXBLOCK_SHIFT))
 
+#define T5577_PAC_CONFIG (        \
+    T5577_MODULATION_DIRECT |     \
+    T5577_BITRATE_RF_32 |         \
+    T5577_PWD |                   \
+    (4 << T5577_MAXBLOCK_SHIFT))
+
 #define T5577_INDALA_64_CONFIG ( \
     T5577_MODULATION_PSK1 |      \
     T5577_BITRATE_RF_32 |        \